+++ punchpipe/flows/level1.py
@@ -1,7 +1,7 @@
 import os
 import json
 import typing as t
-from datetime import datetime
+from datetime import datetime, timedelta
 
 from prefect import flow, get_run_logger, task
 from prefect.cache_policies import NO_CACHE
@@ -31,8 +31,11 @@ def level1_query_ready_files(session, pipeline_config: dict, reference_time=None
         if get_quartic_model_path(f, pipeline_config, session=session) is None:
             logger.info(f"Missing quartic model for {f.filename()}")
             continue
-        if get_stray_light(f, pipeline_config, session=session) is None:
-            logger.info(f"Missing stray light model for {f.filename()}")
+        if get_stray_light_before(f, pipeline_config, session=session) is None:
+            logger.info(f"Missing stray light before model for {f.filename()}")
+            continue
+        if get_stray_light_after(f, pipeline_config, session=session) is None:
+            logger.info(f"Missing stray light after model for {f.filename()}")
             continue
         if get_vignetting_function_path(f, pipeline_config, session=session) is None:
             logger.info(f"Missing vignetting function for {f.filename()}")
@@ -80,7 +83,7 @@ def get_psf_model_path(level0_file, pipeline_config: dict, session=None, referen
                   .order_by(File.date_obs.desc()).first())
     return best_model
 
-def get_stray_light(level0_file, pipeline_config: dict, session=None, reference_time=None):
+def get_stray_light_before(level0_file, pipeline_config: dict, session=None, reference_time=None):
     corresponding_type = {"PM": "SM",
                           "PZ": "SZ",
                           "PP": "SP",
@@ -90,6 +93,22 @@ def get_stray_light(level0_file, pipeline_config: dict, session=None, reference_
                   .filter(File.file_type == model_type)
                   .filter(File.observatory == level0_file.observatory)
                   .where(File.date_obs <= level0_file.date_obs)
+                  .where(File.date_obs > level0_file.date_obs - timedelta(days=1))
+                  .order_by(File.date_obs.desc()).first())
+    return best_model
+
+
+def get_stray_light_after(level0_file, pipeline_config: dict, session=None, reference_time=None):
+    corresponding_type = {"PM": "SM",
+                          "PZ": "SZ",
+                          "PP": "SP",
+                          "CR": "SR"}
+    model_type = corresponding_type[level0_file.file_type]
+    best_model = (session.query(File)
+                  .filter(File.file_type == model_type)
+                  .filter(File.observatory == level0_file.observatory)
+                  .where(File.date_obs >= level0_file.date_obs)
+                  .where(File.date_obs < level0_file.date_obs + timedelta(days=1))
                   .order_by(File.date_obs.desc()).first())
     return best_model
 
@@ -129,7 +148,8 @@ def level1_construct_flow_info(level0_files: list[File], level1_files: File,
     best_quartic_model = get_quartic_model_path(level0_files[0], pipeline_config, session=session)
     best_distortion = get_distortion_path(level0_files[0], pipeline_config, session=session)
     ccd_parameters = get_ccd_parameters(level0_files[0], pipeline_config, session=session)
-    best_stray_light = get_stray_light(level0_files[0], pipeline_config, session=session)
+    best_stray_light_before = get_stray_light_before(level0_files[0], pipeline_config, session=session)
+    best_stray_light_after = get_stray_light_after(level0_files[0], pipeline_config, session=session)
     mask_function = get_mask_file(level0_files[0], pipeline_config, session=session)
 
     call_data = json.dumps(
@@ -148,8 +168,10 @@ def level1_construct_flow_info(level0_files: list[File], level1_files: File,
             "gain_top": ccd_parameters['gain_top'],
             "distortion_path": os.path.join(best_distortion.directory(pipeline_config['root']),
                                             best_distortion.filename()),
-            "stray_light_path": os.path.join(best_stray_light.directory(pipeline_config['root']),
-                                            best_stray_light.filename()),
+            "stray_light_path_before": os.path.join(best_stray_light_before.directory(pipeline_config['root']),
+                                            best_stray_light_before.filename()),
+            "stray_light_path_after": os.path.join(best_stray_light_after.directory(pipeline_config['root']),
+                                            best_stray_light_after.filename()),
             "mask_path": os.path.join(mask_function.directory(pipeline_config['root']),
                                       mask_function.filename().replace('.fits', '.bin')),
             "return_with_stray_light": True,